refactor(style)!: split style.ts, add subdir
Both to better manage a fairly complex single file, and to better insure
cleaner separation of parameter options and template language, split
into multiple files, and move most of them into a style subdirectory.
