+++ src/style/locale.ts
